Limit test rule to test directories and rename base classes
- Add suppression rule to ensure TestClassNamingConvention only applies to test directories
- Rename abstract test base classes (DataTestBase, RecordWrapperTestBase, AvroDataTestBase)
- Update test classes that extend these base classes and reorder import statements
