+++ .baseline/checkstyle/checkstyle-suppressions.xml
@@ -54,4 +54,7 @@
 
     <!-- Suppress checks for CometColumnReader -->
     <suppress files="org.apache.iceberg.spark.data.vectorized.CometColumnReader" checks="IllegalImport"/>
+
+    <!-- Suppress TestClassNamingConvention for main source files -->
+    <suppress files=".*[/\\]src[/\\]main[/\\].*" id="TestClassNamingConvention" />
 </suppressions>
