+++ core/src/test/java/org/apache/iceberg/TestSchemaParser.java
@@ -27,7 +27,7 @@ import java.math.BigDecimal;
 import java.nio.ByteBuffer;
 import java.util.UUID;
 import java.util.stream.Stream;
-import org.apache.iceberg.data.TestData;
+import org.apache.iceberg.data.DataTestBase;
 import org.apache.iceberg.expressions.Literal;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 import org.apache.iceberg.relocated.com.google.common.collect.Sets;
@@ -39,7 +39,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class TestSchemaParser extends TestData {
+public class TestSchemaParser extends DataTestBase {
   @Override
   protected boolean supportsUnknown() {
     return true;
