+++ core/src/test/java/org/apache/iceberg/avro/TestGenericAvro.java
@@ -22,13 +22,13 @@ import java.io.IOException;
 import java.util.List;
 import org.apache.avro.generic.GenericData.Record;
 import org.apache.iceberg.Schema;
-import org.apache.iceberg.data.TestData;
+import org.apache.iceberg.data.DataTestBase;
 import org.apache.iceberg.inmemory.InMemoryOutputFile;
 import org.apache.iceberg.io.FileAppender;
 import org.apache.iceberg.io.OutputFile;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 
-public class TestGenericAvro extends TestData {
+public class TestGenericAvro extends DataTestBase {
   @Override
   protected boolean supportsUnknown() {
     return true;
