+++ core/src/test/java/org/apache/iceberg/avro/TestInternalAvro.java
@@ -24,14 +24,14 @@ import org.apache.iceberg.InternalTestHelpers;
 import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.RandomInternalData;
 import org.apache.iceberg.Schema;
+import org.apache.iceberg.data.DataTestBase;
 import org.apache.iceberg.data.Record;
-import org.apache.iceberg.data.TestData;
 import org.apache.iceberg.inmemory.InMemoryOutputFile;
 import org.apache.iceberg.io.DataWriter;
 import org.apache.iceberg.io.OutputFile;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 
-public class TestInternalAvro extends TestData {
+public class TestInternalAvro extends DataTestBase {
   @Override
   protected boolean supportsDefaultValues() {
     return true;
