+++ data/src/test/java/org/apache/iceberg/RecordWrapperTestBase.java
@@ -26,7 +26,7 @@ import org.apache.iceberg.types.Types;
 import org.apache.iceberg.util.StructLikeWrapper;
 import org.junit.jupiter.api.Test;
 
-public abstract class TestRecordWrapper {
+public abstract class RecordWrapperTestBase {
 
   private static final Types.StructType PRIMITIVE_WITHOUT_TIME =
       Types.StructType.of(
