+++ data/src/test/java/org/apache/iceberg/data/avro/TestGenericData.java
@@ -27,14 +27,14 @@ import org.apache.iceberg.Files;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.avro.Avro;
 import org.apache.iceberg.avro.AvroIterable;
+import org.apache.iceberg.data.DataTestBase;
 import org.apache.iceberg.data.DataTestHelpers;
 import org.apache.iceberg.data.RandomGenericData;
 import org.apache.iceberg.data.Record;
-import org.apache.iceberg.data.TestData;
 import org.apache.iceberg.io.FileAppender;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 
-public class TestGenericData extends TestData {
+public class TestGenericData extends DataTestBase {
 
   @Override
   protected void writeAndValidate(Schema schema) throws IOException {
