+++ spark/v3.4/spark/src/test/java/org/apache/iceberg/spark/source/TestInternalRowWrapper.java
@@ -21,9 +21,9 @@ package org.apache.iceberg.spark.source;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Iterator;
+import org.apache.iceberg.RecordWrapperTestBase;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.StructLike;
-import org.apache.iceberg.TestRecordWrapper;
 import org.apache.iceberg.data.InternalRecordWrapper;
 import org.apache.iceberg.data.RandomGenericData;
 import org.apache.iceberg.data.Record;
@@ -33,7 +33,7 @@ import org.apache.iceberg.util.StructLikeWrapper;
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.junit.jupiter.api.Disabled;
 
-public class TestInternalRowWrapper extends TestRecordWrapper {
+public class TestInternalRowWrapper extends RecordWrapperTestBase {
 
   @Disabled
   @Override
