+++ spark/v3.5/spark/src/test/java/org/apache/iceberg/spark/data/TestAvroData.java
@@ -54,7 +54,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public abstract class TestAvroData {
+public abstract class AvroDataTestBase {
 
   private static final long FIRST_ROW_ID = 2_000L;
   protected static final Map<Integer, Object> ID_TO_CONSTANT =
