+++ spark/v4.0/spark/src/test/java/org/apache/iceberg/spark/source/ScanTestBase.java
@@ -37,8 +37,8 @@ import org.apache.iceberg.TableOperations;
 import org.apache.iceberg.TableProperties;
 import org.apache.iceberg.hadoop.HadoopTables;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
+import org.apache.iceberg.spark.data.AvroDataTestBase;
 import org.apache.iceberg.spark.data.RandomData;
-import org.apache.iceberg.spark.data.TestAvroData;
 import org.apache.iceberg.spark.data.TestHelpers;
 import org.apache.iceberg.types.TypeUtil;
 import org.apache.spark.api.java.JavaSparkContext;
@@ -50,7 +50,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.io.TempDir;
 
 /** An AvroDataScan test that validates data by reading through Spark */
-public abstract class ScanTestBase extends TestAvroData {
+public abstract class ScanTestBase extends AvroDataTestBase {
   private static final Configuration CONF = new Configuration();
 
   protected static SparkSession spark = null;
